Refactor the code structure Add documentation about the features and software
