+++ src/auxiliary_controller/Debug/auxiliary_controller.hex
