+++ src/auxiliary_controller/auxiliary_controller.__w
@@ -1,0 +1,6 @@
+[Workspace]
+Count=0
+Active=
+Proj0=auxiliary_controller.__c
+[Settings]
+Config=1
