+++ src/auxiliary_controller/auxiliary_controller.c
