+++ src/auxiliary_controller/makefile.auxiliary_controller
@@ -1,0 +1,27 @@
+# Generated by SourceBoost IDE 8.01
+
+Debug:
+	@if not exist $@ mkdir $@
+
+CC = "C:\Program Files (x86)\SourceBoost\boostc_pic16.exe"
+
+Debug\auxiliary_controller.obj: auxiliary_controller.c uart.h spi.h adc.h utils.h
+	$(CC) $< -t PIC16F886  -idx 1 -obj Debug -d _DEBUG
+
+LD = "C:\Program Files (x86)\SourceBoost\boostlink_picmicro.exe"
+
+Debug\auxiliary_controller.hex: Debug\auxiliary_controller.obj 
+	$(LD)  -idx 1  /ld "C:\Program Files (x86)\SourceBoost\lib" libc.pic16.lib $+ /t PIC16F886 /d "Debug" /p auxiliary_controller
+
+all: Debug Debug\auxiliary_controller.hex
+
+clean:
+	@if exist Debug\auxiliary_controller.obj del Debug\auxiliary_controller.obj
+	@if exist Debug\auxiliary_controller.hex del Debug\auxiliary_controller.hex
+	@if exist Debug\auxiliary_controller.asm del Debug\auxiliary_controller.asm
+	@if exist Debug\auxiliary_controller.lst del Debug\auxiliary_controller.lst
+	@if exist Debug\auxiliary_controller.map del Debug\auxiliary_controller.map
+	@if exist Debug\auxiliary_controller.cof del Debug\auxiliary_controller.cof
+	@if exist Debug\auxiliary_controller.casm del Debug\auxiliary_controller.casm
+	@if exist Debug\auxiliary_controller.stat del Debug\auxiliary_controller.stat
+	@if exist Debug\auxiliary_controller.tree del Debug\auxiliary_controller.tree
