+++ src/auxiliary_controller/utils.h
