+++ src/components/camera_slave.py
@@ -20,6 +20,77 @@
 #OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 #SOFTWARE.
 
+from pyb import Pin
+import sensor
+
+class CameraSlave():
+
+    control = None
+    rx_image_width = 320
+    rx_image_height = 240
+    spi = None
+    rx_buff = None
+    tx_buff = None
+
+    def __init__(self, spi, width, height, alloc_tx_buff=False, alloc_rx_buff=False, pin_data_ready=None, sync_cancel_condition=None):
+        self.spi = spi
+        if alloc_tx_buff:
+            self.tx_buff = sensor.alloc_extra_fb(width, height, sensor.RGB565)
+        if alloc_rx_buff:
+            self.rx_buff = sensor.alloc_extra_fb(width, height, sensor.RGB565)
+        self.pin_data_ready = Pin(pin_data_ready, Pin.IN)
+        if not sync_cancel_condition:
+            self.sync_cancel_condition = lambda: False
+        else:
+            self.sync_cancel_condition = sync_cancel_condition
+
+        self.control = CameraSlaveControl()
+
+    def sync(self, ignore_busy=False):
+        if not ignore_busy:
+            while self.pin_data_ready and self.pin_data_ready.value() == 1:
+                if self.sync_cancel_condition():
+                    return False
+        if self.tx_buff is not None and self.rx_buff is not None:
+            self.spi.send_recv(self.tx_buff.bytearray(), self.rx_buff.bytearray())
+        elif self.tx_buff is not None:
+            self.spi.send(self.tx_buff.bytearray())
+        elif self.rx_buff is not None:
+            self.spi.recv(self.rx_buff.bytearray())
+
+        self.spi.send(self.control.buff)
+
+        return True
+
+    def increase_column_offset(self):
+        self.control.column_offset += 1
+
+    def decrease_column_offset(self):
+        self.control.column_offset -= 1
+
+    def increase_row_offset(self):
+        self.control.row_offset += 1
+
+    def decrease_row_offset(self):
+        self.control.row_offset -= 1
+
+    def increase_column_factor(self):
+        self.control.column_zoom_numerator += 1
+        self.control.column_zoom_denominator = 20
+
+    def decrease_column_factor(self):
+        self.control.column_zoom_numerator -= 1
+        self.control.column_zoom_denominator = 20
+
+    def increase_row_factor(self):
+        self.control.row_zoom_numerator += 1
+        self.control.row_zoom_denominator = 20
+
+    def decrease_row_factor(self):
+        self.control.row_zoom_numerator -= 1
+        self.control.row_zoom_denominator = 20
+
+
 class CameraSlaveControl():
     COLUMN_OFFSET = const(0)
     ROW_OFFSET = const(1)
