+++ src/openmv_thermal/camera.json
