+++ src/openmv_thermal/startup.bmp
